Move bot out of constants.py to bot.py
